requirements.txt
Flask==1.0.2
flask-bootstrap==3.3.7.1
pymysql==0.9.3
pandas==0.24.2
mysql-connector-python==8.0.15
